+++ src/mol-repr/structure/visual/util/common.ts
@@ -141,7 +141,7 @@ export function getConformation(unit: Unit) {
 
 export const CommonSurfaceParams = {
     ignoreHydrogens: PD.Boolean(false, { description: 'Whether or not to include hydrogen atoms in the surface calculation.' }),
-    onlyPolarHydrogens: PD.Boolean(false, { description: 'Whether or not to include polar hydrogen atoms in the surface calculation.' }),
+    ignoreHydrogensVariant: PD.Select('all', PD.arrayToOptions(['all', 'non-polar'] as const)),
     traceOnly: PD.Boolean(false, { description: 'Whether or not to only use trace atoms in the surface calculation.' }),
     includeParent: PD.Boolean(false, { description: 'Include elements of the parent structure in surface calculation to get a surface patch of the current structure.' }),
 };
@@ -169,7 +169,7 @@ function filterUnitId(id: AssignableArrayLike<number>, elements: SortedArray, in
 }
 
 export function getUnitConformationAndRadius(structure: Structure, unit: Unit, sizeTheme: SizeTheme<any>, props: CommonSurfaceProps) {
-    const { ignoreHydrogens, onlyPolarHydrogens, traceOnly, includeParent } = props;
+    const { ignoreHydrogens, ignoreHydrogensVariant, traceOnly, includeParent } = props;
     const rootUnit = includeParent ? structure.root.unitMap.get(unit.id) : unit;
     const differentRoot = includeParent && rootUnit !== unit;
 
@@ -182,13 +182,12 @@ export function getUnitConformationAndRadius(structure: Structure, unit: Unit, s
     let indices: SortedArray<ElementIndex>;
     let id: AssignableArrayLike<number>;
 
-    if (ignoreHydrogens || onlyPolarHydrogens || traceOnly || differentRoot) {
+    if (ignoreHydrogens || traceOnly || differentRoot) {
         const _indices: number[] = [];
         const _id: number[] = [];
         for (let i = 0, il = elements.length; i < il; ++i) {
             const eI = elements[i];
-            if (ignoreHydrogens && isHydrogen(rootUnit, eI)) continue;
-            if (onlyPolarHydrogens && isHydrogen(rootUnit, eI) && Unit.isAtomic(rootUnit) && !hasPolarNeighbour(structure, rootUnit, SortedArray.indexOf(rootUnit.elements, eI) as StructureElement.UnitIndex)) continue;
+            if (ignoreHydrogens && isHydrogen(structure, rootUnit, eI, ignoreHydrogensVariant)) continue;
             if (traceOnly && !isTrace(rootUnit, eI)) continue;
             if (differentRoot && squaredDistance(x[eI], y[eI], z[eI], center) > radiusSq) continue;
 
@@ -219,7 +218,7 @@ export function getUnitConformationAndRadius(structure: Structure, unit: Unit, s
 }
 
 export function getStructureConformationAndRadius(structure: Structure, sizeTheme: SizeTheme<any>, props: CommonSurfaceProps) {
-    const { ignoreHydrogens, onlyPolarHydrogens, traceOnly, includeParent } = props;
+    const { ignoreHydrogens, ignoreHydrogensVariant, traceOnly, includeParent } = props;
     const differentRoot = includeParent && !!structure.parent;
     const l = StructureElement.Location.create(structure.root);
 
@@ -234,7 +233,7 @@ export function getStructureConformationAndRadius(structure: Structure, sizeThem
     let id: AssignableArrayLike<number>;
     let indices: OrderedSet<number>;
 
-    if (ignoreHydrogens || onlyPolarHydrogens || traceOnly || differentRoot) {
+    if (ignoreHydrogens || traceOnly || differentRoot) {
         const { getSerialIndex } = structure.serialMapping;
         const units = differentRoot ? structure.root.units : structure.units;
 
@@ -252,8 +251,7 @@ export function getStructureConformationAndRadius(structure: Structure, sizeThem
             l.unit = unit;
             for (let j = 0, jl = elements.length; j < jl; ++j) {
                 const eI = elements[j];
-                if (ignoreHydrogens && isHydrogen(unit, eI)) continue;
-                if (onlyPolarHydrogens && isHydrogen(unit, eI) && Unit.isAtomic(unit) && !hasPolarNeighbour(structure, unit, SortedArray.indexOf(unit.elements, eI) as StructureElement.UnitIndex)) continue;
+                if (ignoreHydrogens && isHydrogen(structure, unit, eI, ignoreHydrogensVariant)) continue;
                 if (traceOnly && !isTrace(unit, eI)) continue;
 
                 const _x = x(eI), _y = y(eI), _z = z(eI);
@@ -316,9 +314,12 @@ export function getStructureConformationAndRadius(structure: Structure, sizeThem
 }
 
 const _H = AtomicNumbers['H'];
-export function isHydrogen(unit: Unit, element: ElementIndex) {
+export function isHydrogen(structure: Structure, unit: Unit, element: ElementIndex, variant: 'all' | 'non-polar') {
     if (Unit.isCoarse(unit)) return false;
-    return unit.model.atomicHierarchy.derived.atom.atomicNumber[element] === _H;
+    return (
+        unit.model.atomicHierarchy.derived.atom.atomicNumber[element] === _H &&
+        (variant === 'all' || !hasPolarNeighbour(structure, unit, SortedArray.indexOf(unit.elements, element) as StructureElement.UnitIndex))
+    );
 }
 export function isH(atomicNumber: ArrayLike<number>, element: ElementIndex) {
     return atomicNumber[element] === _H;